Add Declarative Pipeline to Jenkinsfile
Added Declarative Pipeline to Jenkinsfile

`agent` section with `none` parameter means that no global
agent will be allocated for the entire Pipeline's execution
and that each `stage` directive must specify its own `agent`
section

`stage('Build')` defines a `stage` directive `Build`
that appears on the Jenkins UI.

`image` parameter downloads the corresponding Docker image
(if not already available on the machine) and runs it as a
separate container.

The Python container will be the `agent` that Jenkins uses
to run the `Build` stage of this Pipeline. Its lifespan
is only that of the duration of your Build stage's execution.

The `sh` step (of the steps section) runs the Python command
to compile the application and its `calc` library into byte code files
which are placed into the sources workspace directory
(within the /var/jenkins_home/workspace/simple-python-pyinstaller-app directory
in the Jenkins container).