Add application Jenkinsfile
Added application Jenkinsfile with Declarative Pipeline.

`agent` section with the `none` parameter specified means
that no global `agent` will be allocated for the entire
Pipeline’s execution and that each stage directive must
specify its own `agent` section.

`stage('Build')` defines a directive called `Build` that
appears on the Jenkins UI.
The `image` parameter (of the `agent` section’s `docker` parameter)
downloads the python:2-alpine Docker image (if it’s not already
available locally) and runs this image as a separate container.

The Python container becomes the agent that Jenkins uses to run
the `Build` stage of the Pipeline project.
The container is short-lived - its lifespan is only that of
the duration of your Build stage’s execution.

The `sh` step (of the steps section) runs the Python command to
compile the application and its `calc` library into byte code files
(with .pyc extension), which are placed into the sources workspace
directory
(within the /var/jenkins_home/workspace/simple-python-pyinstaller-app
directory in the Jenkins container).