Add a test stage to the Pipeline
Added a `stage` directive called `Test` that appears
on the Jenkins UI.

Added the `image` parameter that downloads the `qnib:pytest`
Docker image (if not available locally) and runs it as a separate
container. The `pytest` container becomes the agent that Jenkins
uses to run the `Test` stage of your Pipeline project.

The `pytest` container's lifespan lasts the duration of
your `Test` stage's execution.